resolveStepLayout function
